pollution_backend/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Machine Learning
tensorflow==2.20.0
numpy==1.24.3
Pillow==10.0.1
scikit-learn==1.3.0

# Image processing
opencv-python==4.8.1.78

# Database
SQLAlchemy==2.0.19
psycopg2-binary==2.9.7
PyMySQL==1.1.0

# Satellite imagery
earthengine-api==0.1.350
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development and testing
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.0.0

# Production
gunicorn==21.2.0
